ugc/src/requirements.txt
aiokafka==0.7.2
anyio==3.6.1
cffi==1.15.1
click==8.1.3
clickhouse-driver==0.2.4
colorama==0.4.5
cryptography==38.0.1
fastapi==0.85.1
fastapi-jwt-auth==0.5.0
sniffio==1.3.0
starlette==0.20.4
typing-extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
uvicorn==0.18.3
win32-setctime==1.1.0
gunicorn==20.1.0
confluent-kafka==1.9.2
pydantic[dotenv]
httpx==0.22.0